Breakout shard query logic into sharded base query logic to support different return types for upcoming JSON PR
