+++ warehouse/query-core/src/main/java/datawave/query/config/Profile.java
@@ -12,31 +12,31 @@ import org.apache.accumulo.core.data.Value;
 /**
  * Purpose: Provides a mechanism to provide a user configurable way to tune a user's query, such that he or she may take advantage of features that could have
  * negative implications on other user queries.
- * 
+ *
  * Design: Simple interface. User would select the profile through a bean
- * 
+ *
  *
  */
-public interface Profile {
+public interface Profile<T> {
     /**
      * Tune the query logic
-     * 
+     *
      * @param logic
      *            the query logic
      */
-    void configure(BaseQueryLogic<Entry<Key,Value>> logic);
-    
+    void configure(BaseQueryLogic<T> logic);
+
     /**
      * Tune the query planner.
-     * 
+     *
      * @param planner
      *            the query planner.
      */
     void configure(QueryPlanner planner);
-    
+
     /**
      * Tune the query configuration object
-     * 
+     *
      * @param configuration
      *            the query configuration object
      */
